requirements.txt
flask
opencv-python-headless
Pillow
numpy
torch
diffusers
transformers
accelerate
safetensors
